+++ examples/backend.tf
@@ -1,6 +1,6 @@
 module "backendifier" {
   source  = "gekken/backendifier/aws"
-  version = "1.1.0"
+  version = "1.1.1"
 
   # Team decided that "tf_backend" is too boring!
   bucket_name_suffix = "my-super-duper-project"
@@ -8,5 +8,5 @@ module "backendifier" {
 
 output "where_is_my_bucket" {
   description = "Where, precisely, IS by bucket?"
-  value       = module.backend.bucket_name
+  value       = module.backendifier.bucket_name
 }