+++ s3.tf
@@ -1,10 +1,17 @@
 # State Bucket
 
 resource "aws_s3_bucket" "state_bucket" {
-  #checkov:skip=CKV2_AWS_62:Handled in TF/Pipeline
   #checkov:skip=CKV2_AWS_61:Deprecated, Handled in aws_s3_bucket_lifecycle_configuration
-  #checkov:skip=CKV_AWS_144: TODO: handle XRR
-  bucket = local.bucket_name
+  #checkov:skip=CKV2_AWS_62:Handled in TF/Pipeline
+  #checkov:skip=CKV_AWS_144:Don't see the point for terraform. if you want it, make more buckets in EU or something.
+  bucket = var.custom_bucket_name != null ? var.custom_bucket_name : local.bucket_name
+
+  tags = var.tagging
+}
+
+resource "aws_s3_bucket_policy" "s3_bucket_force_ssl" {
+  bucket = aws_s3_bucket.state_bucket.id
+  policy = data.aws_iam_policy_document.s3_bucket_force_ssl.json
 }
 
 resource "aws_s3_bucket_versioning" "state_bucket_versioning" {
@@ -16,9 +23,15 @@ resource "aws_s3_bucket_versioning" "state_bucket_versioning" {
 }
 
 resource "aws_s3_bucket_lifecycle_configuration" "state_bucket_lifecycle" {
-  #checkov:skip=CKV_AWS_300:TODO:Handle failed uploads?
   bucket = aws_s3_bucket.state_bucket.id
+  rule {
+    id     = "abort-failed-uploads"
+    status = "Enabled"
 
+    abort_incomplete_multipart_upload {
+      days_after_initiation = 1
+    }
+  }
   rule {
     id     = "auto-archive"
     status = "Enabled"
@@ -50,17 +63,13 @@ resource "aws_s3_bucket_logging" "state_bucket_logging" {
 }
 
 resource "aws_s3_bucket_ownership_controls" "state_bucket_ownership" {
+  #checkov:skip=CKV2_AWS_65:This is unnecessary with IAM, no public, etc. what about several teams using one bucket in an Org?
   bucket = aws_s3_bucket.state_bucket.id
 
   rule {
     object_ownership = "BucketOwnerPreferred"
   }
 }
-
-# resource "aws_s3_bucket_acl" "state_bucket_acls" {
-#   bucket = aws_s3_bucket.state_bucket.id
-#   acl    = "private"
-# }
 
 resource "aws_s3_bucket_public_access_block" "state_bucket_public_access_blocking" {
   bucket = aws_s3_bucket.state_bucket.id
@@ -89,22 +98,20 @@ resource "aws_s3_bucket" "state_logging_bucket" {
   #checkov:skip=CKV_AWS_21:Logging Bucket versioning is unnecessary for logging bucket
   #checkov:skip=CKV2_AWS_61:Deprecated, Handled in aws_s3_bucket_lifecycle_configuration
   #checkov:skip=CKV2_AWS_62:Handled by pipeline/terraform
-  #checkov:skip=CKV_AWS_144: TODO: handle XRR
-  bucket = "${local.bucket_name}-logging"
+  #checkov:skip=CKV_AWS_144:No point in terraform.
+  bucket = "${aws_s3_bucket.state_bucket.id}-logging"
+
+  tags = var.tagging
 }
 
 resource "aws_s3_bucket_ownership_controls" "state_logging_bucket_ownership" {
+  #checkov:skip=CKV2_AWS_65:This is to handle multi-account Orgs.
   bucket = aws_s3_bucket.state_logging_bucket.id
 
   rule {
     object_ownership = "BucketOwnerPreferred"
   }
 }
-
-# resource "aws_s3_bucket_acl" "state_logging_bucket_acls" {
-#   bucket = aws_s3_bucket.state_logging_bucket.id
-#   acl    = "private"
-# }
 
 resource "aws_s3_bucket_public_access_block" "state_logging_bucket_public_access_blocking" {
   bucket = aws_s3_bucket.state_logging_bucket.id
@@ -127,7 +134,7 @@ resource "aws_s3_bucket_server_side_encryption_configuration" "state_logging_buc
 }
 
 resource "aws_s3_bucket_lifecycle_configuration" "state_logging_bucket_lifecycle" {
-  #checkov:skip=CKV_AWS_300:TODO:Handle failed uploads?
+  #checkov:skip=CKV_AWS_300:If AWS is failing their own uploads, there are other issues.
   bucket = aws_s3_bucket.state_logging_bucket.id
 
   rule {